replace sea-orm with sqlx postgres only
