+++ Cargo.toml
@@ -29,8 +29,6 @@ once_cell = { version = "^1.17" }
 reqwest = { version = "^0.11", optional = true, default-features = false, features = ["blocking"] }
 rust-embed = { version = "^6.6.1", features = ["salvo-ex"] }
 salvo = { version = "0.41", features = ["affix", "anyhow", "catch-panic", "cors", "proxy", "logging", "serve-static", "affix", "size-limiter", "oapi"] }
-sea-orm = { version = "^0.11", features = ["debug-print", "sqlx-postgres", "runtime-tokio-rustls", "macros"] }
-sea-orm-migration = { version = "0.11.3", features = ["runtime-tokio-rustls", "sea-orm-cli", "sqlx-postgres"] }
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = { version = "^1.0" }
 sqlx = { version = "^0.6", features = [ "runtime-tokio-rustls", "migrate", "postgres", "uuid", "chrono", "json" ] }
