+++ src/handler/admin.rs
@@ -1,8 +1,5 @@
 // Copyright 2022-current Aris Ripandi <aris@duck.com>
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod app;
-pub mod database;
-
-pub use self::app::*;
-pub use self::database::*;
+// get_all_users
+// invite_by_admin
