--- src/handler/admin/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2022-current Aris Ripandi <aris@duck.com>
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod get_all_users;
-pub mod invite_by_admin;
-
-pub use self::get_all_users::*;
-pub use self::invite_by_admin::*;
