--- src/handler/auth/mod.rs
@@ -1,14 +1,0 @@
-// Copyright 2022-current Aris Ripandi <aris@duck.com>
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod logout;
-pub mod recover;
-pub mod signup;
-pub mod token;
-pub mod verify;
-
-pub use self::logout::*;
-pub use self::recover::*;
-pub use self::signup::*;
-pub use self::token::*;
-pub use self::verify::*;
