--- src/handler/auth/verify.rs
@@ -1,2 +1,0 @@
-// Copyright 2022-current Aris Ripandi <aris@duck.com>
-// SPDX-License-Identifier: Apache-2.0
