+++ src/handler/logout.rs
