--- src/utils/validator/mod.rs
@@ -1,6 +1,0 @@
-// Copyright 2022-current Aris Ripandi <aris@duck.com>
-// SPDX-License-Identifier: Apache-2.0
-
-mod is_valid_email;
-
-pub use self::is_valid_email::*;
